+++ Tests/NotchIslandTests/IslandLayoutTests.swift
@@ -1,5 +1,6 @@
 import CoreGraphics
 import Testing
+
 @testable import NotchIsland
 
 @Suite("Island geometry")