+++ Tests/NotchIslandTests/LineAccumulatorTests.swift
@@ -1,5 +1,6 @@
 import Foundation
 import Testing
+
 @testable import NotchIsland
 
 @Suite("Line reassembly")